+++ webpack.config.dist.js
@@ -44,10 +44,6 @@ module.exports = {
             }
           ]
         })
-      },
-      {
-        test: /\.md$/,
-        loader: 'raw-loader'
       }
     ]
   },