+++ webpack.config.js
@@ -48,10 +48,6 @@ module.exports = {
             }
           ]
         })
-      },
-      {
-        test: /\.md$/,
-        loader: 'raw-loader'
       }
     ]
   },